+++ metapredict/metapredict_exceptions.py
@@ -6,3 +6,4 @@ class DomainError(Exception):
 
 class MetapredictError(Exception):
     pass
+